refactor(hogql): more cleanup
